+++ NEWS.md
@@ -1,8 +1,8 @@
 # mini007 0.2.0
 
+- Setting `messages` as an `active` `R6` field. It is now possible to modify the `messages` that will be used by the `LLM` object through a `list` object. The `list` object will be automatically convert to the corresponding `ellmer` `Turns`. 
 
 Adding the following new methods: 
-
 - `truncate_history()`